Task lww 56 deposit fiat using credit card fx payment gateway (#85)
* LWDEV-5715 affiliate client service added

* LWDEV-5715 Affiliate Client updated to 1.0.6

* LWDEV-5715 add statistics for affiliate program

* LWDEV-5715 stats method updated

* LWDEV-5715 affiliate methods divided

* LWDEV-5715 affiliate - GetLink updated to return NotFound

* added fees to the history response

* added new fields to fee for limit order

* LWDEV-5561 session validation

* temporarily removed singfilter for testing

* wrong phone sign filter commented

* PhoneSignFilter returned

* LWW-10 Implementation of Settings controller and needed functionality

* PhoneSignFilter removed for LO testing

* PhoneSignFilter commented everywhere

* PhoneSignFilter for market order

* LWW-22 Start reimplementation

* LWDEV-5561 session validation

* LWDEV-5561 session validation

* LWDEV-5561 session validation

* LWW-22 Fix references

* LWW-22 Code refactoring

* LWW-22 Fix tests

* LWW-24 Implementing CountryPhoneCodesController, models and services

* LWDEV-5561

* LWW-24 Add filter registration

* nugets updated

* temp method removed

* ME connector updated

* nugets updated and invalid price handling added

* LWW-26 Reworked solution

* LWW-26 Refactoring

* Fix model name

* LWW-26 Some refactoring fixes after merge

* LWW-26 Remove whitespaces and empty lines

* LWW-27 Refactor endpoint

* session confirmation removed

* LWW-7 GenerateNewIdAsync was fixed

* LWDEV-5956 removed catching error, let lykke middleware to log it

* LWW-27 Refactor endpoints and project

* LWW-27 Removed comment

* LWW-27 Refactoring

* LWW-7 Repair Dependency of service

* LWDEV-6587 place market order

* LWDEV-6587 place limit order

* updated ME connector, pass fee in Fees property in limit order

* LWW-27 Return "using" back

* LWW-10 Implementation of Settings controller and needed functionality

* LWW-22 Start reimplementation

* LWW-22 Fix references

* LWW-22 Code refactoring

* LWW-22 Fix tests

* LWW-24 Implementing CountryPhoneCodesController, models and services

* LWW-24 Add filter registration

* LWW-26 Reworked solution

* LWW-26 Refactoring

* Fix model name

* LWW-26 Some refactoring fixes after merge

* LWW-26 Remove whitespaces and empty lines

* LWW-27 Refactor endpoint

* LWW-7 GenerateNewIdAsync was fixed

* LWW-27 Refactor endpoints and project

* LWW-27 Removed comment

* LWW-27 Refactoring

* LWW-7 Repair Dependency of service

* LWW-27 Return "using" back

* LWW-7 Refactoring and fixing after rebase

* LWW-7 Fix reference

* LWDEV-6587 place order

* LWW-7 Refactoring

* LWW-7 Refactoring

* LWW-7 Refactored and added KycStatus

* LWW-7 Fix

* lykke.logs up to date

* LWW-7 Move repositories to services

* LWW-7 Add Endpoint for ConfigData

* LWW-7 Fix endpoint

* LWDEV-5561 session validation

* EnableSessionValidation = false (temporary)

* EnableSessionValidation = false (temporary)

* check for password complexity

* changed password validation on login

* removed pasword length check on login

* LWW-7 Rework model in accordance with approved requirements

* LWW-7 Refactoring

* LWW-7 Refactor FeeSettingsController

* LWW-7 Refactoring

* additional checks added

* LWDEV-6587 404 fix

* removed check for password complexity

* confirmation disabled

* LWDEV-5561 session validation

* LWDEV-5561 session validation

* LWDEV-5561 session validation

* sessions fix

* LWDEV-5561 trading session ttl in ms

* LWDEV-5561 trading session ttl in ms

* LWW-30 The logic is switched to using Lykke.Service.PaymentSystem micro-service.

* LWDEV-7041 kyc status

* assets check to balances added

* LWDEV-5561 EnableSessionValidation flag

* LWDEV-5561 session validation

* LWW-30 PaymentSystemsSettings model was removed
PaymentSystem logic was changed

* LWW-30 Bug with GetUrlDataAsync was fixed

* refactoring

* logout added

* sessions nuget updated

* Revert "Merge branch 'feature_LWDEV-7148_logout' into dev"

This reverts commit be42a841b4ca8317984ba72a63666dd27cd01e06, reversing
changes made to 6e281dc25fed3c7990d5a03206ff688ec71bc83b.

* Revert "Revert "Merge branch 'feature_LWDEV-7148_logout' into dev""

This reverts commit ba7632b016683adbe9b9978d91e6c71879987ca4.

* Revert "Revert "Revert "Merge branch 'feature_LWDEV-7148_logout' into dev"""

This reverts commit c88578e107272c2efa593811698a44737011356c.

* LWDEV-7041: map kyc statuses

* LWW-72 Is not approved deposit disclaimers validation was added

* payment endpoint added

* payment fixes

* reference fix

* minor payment fix

* payment improvements

* remove duplicated reference

* LWW-72 AssetId validation was added

* LWW-72 Validation for BankCardPaymentUrl was added, but doesn't tested

* LWW-72 Bugs were waxed

* LWW-56 Fix after merge

* LWW-30 BankCardPaymentUrl was moved to PaymentSystemService

* LWW-56 <Version> was added to AzureRepositories

* LWW-56 DictionaryController was fixed

* LWW-98 Refactoring for new SettingsService

* LWW-72 IsOtherDepositOptionsEnabled and IsBankCardDepositOptionsEnabled was added

* LWW-112 remove enpoint and unused controller

* LWW-112 removed file from

* LWW-117 PaymentMethods endpoint was added

* LWW-117 Authorize was added

* LWW-117 Payment methods validation was added

* LWW-72: ErrorMessage was added for IsValidLimitation

* LWW-72 LimitIsExceeded message was added

* LWW-119 Validation was added

* LWW-119 Phone validation was fixed

* LWW-56 refactoring

+++ LykkeApi2/Models/ClientBalancesModels/ClientBalanceResponseModel.cs
@@ -5,8 +5,8 @@ namespace LykkeApi2.Models.ClientBalancesModels
     public class ClientBalanceResponseModel
     {
         public string AssetId { get; set; }
-        public double? Balance { get; set; }
-        public double? Reserved { get; set; }
+        public decimal? Balance { get; set; }
+        public decimal? Reserved { get; set; }
 
         public static ClientBalanceResponseModel Create(ClientBalanceModel src)
         {
